% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReacTranScripter-package.R
\docType{package}
\name{ReacTranScripter-package}
\alias{ReacTranScripter}
\alias{ReacTranScripter-package}
\title{ReacTranScripter: Reactive Transport Model Scripter}
\description{
Create Reactive Transport Models from configuration files. The resulting scripts are in the style of the ReacTran R-package, using the package functionality. Equivalent models can be scripted in R and Fortran code.
}
\author{
\strong{Maintainer}: Robert Kotschenreuther \email{robert.kotschenreuther@gmx.de} (\href{https://orcid.org/0009-0001-0375-9579}{ORCID})

}
\keyword{internal}
